WIP: Create Tx/Rx threads once on the echotimer
Instead of creating a new pair of Tx/Rx threads on every call to work,
create a pair of threads once on the constructor. Keep these threads
alive throughout the block's existence and put them to sleep in between
work calls. Use condition variables to signal when they have work to do.