Control in GPIO wait value and out GPIO Tx value
Before this patch, the burst Tx block would wait until the input GPIO is
high and set the out GPIO to high on Tx. With this patch, the wait value
and output values can be controlled. For instance, the block can wait
until the input value becomes low, and set the output GPIO to low when
transmitting and high otherwise.