Add option to wait input GPIO transition
In addition to waiting until the input GPIO changes to a specific value,
add option to expect also that the target value is reached after a
transition. In this case, the first reading shall fetch a value that is
not the target yet, but later reads eventually should obtain the target
value before the timeout.

This feature is useful if the input GPIO brings a confirmation of a
change request via the output GPIO pin. If there is a delay until the
change takes effect (e.g., an electromechanical T/R switch), the first
readings could still fetch the old value before the change is applied.
Hence, by expecting a transition, we avoid a premature confirmation
based on the old value.